+++ tests/remote-cache/test.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-# Note: Most of this test runs as Earthly-in-Earthly so that we can easily send local cache to a tmpfs.
+# Note: Most of this test runs as Earthly-in-Earthly so that we can easily send local cache to a tmpfs; however it
+# must be started outside of earthly.
+#
+# To run this test directly: ./test.sh --build-arg DOCKERHUB_AUTH=true --build-arg DOCKERHUB_USER_SECRET=+secrets/earthly-technologies/dockerhub-mirror/user --build-arg DOCKERHUB_TOKEN_SECRET=+secrets/earthly-technologies/dockerhub-mirror/pass --build-arg DOCKERHUB_MIRROR=registry-1.docker.io.mirror.corp.earthly.dev
 
 set -uxe
 set -o pipefail